+++ R/TAF-package.R
@@ -7,8 +7,6 @@
 #' @title Transparent Assessment Framework for Reproducible Research
 #'
 #' @description
-
-
 #' Functions to organize data, methods, and results used in scientific analyses.
 #' A TAF analysis consists of four scripts (`data.R`, `model.R`, `output.R`,
 #' `report.R`) that are run sequentially. Each script starts by reading files
@@ -16,7 +14,8 @@
 #'
 #' Convenience functions are provided to version control the required data and
 #' software, run analyses, clean residues from previous runs, manage files,
-#' manipulate tables, and produce figures.
+#' manipulate tables, and produce figures. With a focus on stability and
+#' reproducible analyses, TAF is designed to have no package dependencies.
 #'
 #' TAF forms a base layer for the
 #' [icesTAF](https://cran.r-project.org/package=icesTAF) package and other
@@ -34,6 +33,7 @@
 #' \emph{Running scripts:}
 #' \tabular{ll}{
 #'   \code{\link{clean}}      \tab clean TAF directories\cr
+#'   \code{\link{clean.boot}} \tab clean boot directory\cr
 #'   \code{\link{make}}       \tab run R script if needed\cr
 #'   \code{\link{make.all}}   \tab run all TAF scripts as needed\cr
 #'   \code{\link{make.taf}}   \tab run TAF script if needed\cr