+++ R/taf.skeleton.R
@@ -7,6 +7,8 @@
 #' @param force whether to overwrite existing scripts.
 #' @param pkgs packages to load at the start of each script. The default is the
 #'        TAF package, i.e. \code{library(TAF)}.
+#' @param model.script model script filename, either \code{model.R} (default) or
+#'        \code{method.R}.
 #'
 #' @return Full path to analysis directory.
 #'
@@ -22,7 +24,8 @@
 #'
 #' @export
 
-taf.skeleton <- function(path = ".", force = FALSE, pkgs = "TAF")
+taf.skeleton <- function(path = ".", force = FALSE, pkgs = "TAF",
+                         model.script = "model.R")
 {
   # only overwrite files if force = TRUE
   safe.cat <- function(..., file, force) {
@@ -37,20 +40,20 @@ taf.skeleton <- function(path = ".", force = FALSE, pkgs = "TAF")
   on.exit(setwd(owd))
 
   # create initial directories
-  mkdir("bootstrap/initial/data")
+  mkdir("boot/initial/data")
 
   # define headers
-  template <-
-    paste0(
-      "## %s\n\n## Before:\n## After:\n\n",
-      paste("library(", pkgs, ")", collapse = "\n", sep = ""),
-      "\n\nmkdir(\"%s\")\n\n"
-    )
+  template <- paste0("## %s\n\n## Before:\n## After:\n\n",
+                     paste0("library(", pkgs, ")", collapse = "\n"),
+                     "\n\nmkdir(\"%s\")\n\n")
   headers <- list(
     data = "Preprocess data, write TAF data tables",
     model = "Run analysis, write model results",
     output = "Extract results of interest, write TAF output tables",
     report = "Prepare plots and tables for report")
+  if (model.script %in% c("method", "method.R")) {
+    names(headers)[2] <- "method"
+  }
 
   # create TAF scripts
   for (section in names(headers)) {